Get rid of rhs field in variable declaration node
The variable declaration IR node no longer has an optional argument for
an assignment. This was previously there to allow for compound
declaration-assignment statements. We wanted this for intermediate
variables, but no longer require this, since at the IR level
intermediate variables are typed. In the IR these can just be separate
and we can detect unassigned variables.