Add more traits to IR types
Added several more traits to IR types exposed as member functions.
Currently, the type checker uses both the actual type of the type, as
well as functions on the type instance to infer information about the
type. For example, if we have

    t = QuantitativeType(...)

we may do both of the following:

    if isinstance(t, InputType):
        ...
    if t.is_nominal():
        ...

This mixed paradigm is not clean and is subject to errors. It would be
much better to merge all of this in either direction. Leaving this as a
TODO.